+++ src/components/MultiSelect/MultiSelect.test.tsx
@@ -263,6 +263,34 @@ describe.skip("MultiSelect", () => {
     ).not.toBeInTheDocument();
   });
 
+  it("closes the multiselect when the 'esc' key is pressed", async () => {
+    render(
+      <MultiSelect
+        id="multiselect-test-id"
+        buttonText="Multiselect button text"
+        defaultItemsVisible={defaultItemsVisible}
+        items={items}
+        isDefaultOpen={true}
+        isSearchable={false}
+        isBlockElement={false}
+        selectedItems={selectedTestItems}
+        onChange={() => null}
+        onClear={() => null}
+      />
+    );
+
+    const multiSelectButton = screen.getByRole("button");
+    const multiSelectCheckbox = screen.getByRole("checkbox", { name: /dogs/i });
+
+    expect(multiSelectButton.getAttribute("aria-expanded")).toEqual("true");
+
+    await userEvent.click(multiSelectCheckbox);
+
+    await userEvent.keyboard("[Escape]");
+
+    expect(multiSelectButton).toHaveAttribute("aria-expanded", "false");
+  });
+
   it("should allow user to toggle menu by clicking menu button or use the 'Enter'/'Spacebar' key", () => {
     render(
       <MultiSelect