+++ CHANGELOG.md
@@ -10,11 +10,25 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
-- Adds Storybook interaction tests for `Searchbar`.
+- Adds Storybook interaction tests for `CheckboxGroup`, `DatePicker`, and `Slider` components.
+
+### Updates
+
+- Updates base font stack to only `system-ui` and `sans-serif`.
+
+## 3.1.6 (June 20, 2024)
+
+### Adds
+
+- Adds Storybook interaction tests for `Searchbar`, `Select`, and `Radiogroup` components.
 
 ### Updates
 
 - Updates how links are styled within the `Heading` component.
+
+### Fixes
+
+- Fixes broken Github Action for release tags.
 
 ## 3.1.5 (June 6, 2024)
 
@@ -42,6 +56,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Exports `useMediaQuery` and `chakra` from Chakra UI.
 - Updates `Accordion` to close focused panel when "esc" key is pressed.
 - Updates the `TextInput` and `SearchBar` components to better associate the input element to the entire component's helper text.
+- Updates `Accordion` to close panel when element within panel is focused and "esc" key is pressed.
 
 ## 3.1.4 (May 23, 2024)